FIX : make the heading appear in the satellite tracker FIX : make the GPS dock smaller
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@10539 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ GPS/SatelliteStrengthView.h
@@ -9,9 +9,9 @@ class Satellite
 {
 	public:
 		int Id;
-		float Azimuth;
-		float Elevation;
-		float SignalStrength;
+		int Azimuth;
+		int Elevation;
+		int SignalStrength;
 };
 
 void sortSatellitesById(std::vector<Satellite>& List);
